# do the division with a LUT