Recover the optimization that we already had in puppi before the code reorganization.
Saves a factor 2 of latency (57 -> 24 cycles), a factor 10 of LUTs and a factor 5 in FFs for the price  of 63 more BRAMs (2.1% of the number available in a VX690T)

## After the change
````
    +-----+-----+-----+-----+----------+
    |  Latency  |  Interval | Pipeline |
    | min | max | min | max |   Type   |
    +-----+-----+-----+-----+----------+
    |   24|   24|    1|    1| function |
    +-----+-----+-----+-----+----------+

+-----------------+---------+-------+--------+--------+
|       Name      | BRAM_18K| DSP48E|   FF   |   LUT  |
+-----------------+---------+-------+--------+--------+
|DSP              |        -|     16|       -|       -|
|Expression       |        -|      -|       0|    3051|
|FIFO             |        -|      -|       -|       -|
|Instance         |       63|    189|    6615|    1260|
|Memory           |        5|      -|       0|       0|
|Multiplexer      |        -|      -|       -|    4320|
|Register         |        -|      -|   12116|     667|
+-----------------+---------+-------+--------+--------+
|Total            |       68|    205|   18731|    9298|
+-----------------+---------+-------+--------+--------+

````

## Before

````
    +-----+-----+-----+-----+----------+
    |  Latency  |  Interval | Pipeline |
    | min | max | min | max |   Type   |
    +-----+-----+-----+-----+----------+
    |   57|   57|    1|    1| function |
    +-----+-----+-----+-----+----------+

+-----------------+---------+-------+--------+--------+
|       Name      | BRAM_18K| DSP48E|   FF   |   LUT  |
+-----------------+---------+-------+--------+--------+
|DSP              |        -|     72|       -|       -|
|Expression       |        -|      -|       0|    3481|
|FIFO             |        -|      -|       -|       -|
|Instance         |        -|    126|  100170|   98280|
|Memory           |        5|      -|       0|       0|
|Multiplexer      |        -|      -|       -|     936|
|Register         |        -|      -|   12015|    1457|
+-----------------+---------+-------+--------+--------+
|Total            |        5|    198|  112185|  104154|
+-----------------+---------+-------+--------+--------+
|Available        |     2940|   3600|  866400|  433200|
+-----------------+---------+-------+--------+--------+

````